moved the window declaration in EE file as its dependency will be updated in EE
