+++ app/client/src/ce/configs/types.ts
@@ -16,7 +16,7 @@ export type FeatureFlagConfig = {
   default: FeatureFlags;
 };
 
-export type AppsmithUIConfigs = {
+export interface AppsmithUIConfigs {
   sentry: {
     enabled: boolean;
     dsn: string;
@@ -78,4 +78,4 @@ export type AppsmithUIConfigs = {
     apiKey: string;
   };
   appsmithSupportEmail: string;
-};
+}
